pyproject.toml
[project]
name = "bible-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "click>=8.0",
    "faiss-cpu>=1.7",
    "neo4j>=5.0",
    "numpy>=1.21",
    "sentence-transformers>=2.2",
    "tqdm>=4.6",
]
